requirements.txt
flask==2.1.3
SQLAlchemy==1.4
babel==2.9.0
python-dateutil==2.6.0
flask-moment==0.11.0
flask-wtf==0.14.3
flask_sqlalchemy==2.4.4
SQLAlchemy
python-dotenv
Werkzeug==2.0.0
jinja2==3.0.1
MarkupSafe==2.1.1